+++ packages/stacks/firebase/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @elbwalker/stack-firebase
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [d17e25a]
+- Updated dependencies [451e1ae]
+- Updated dependencies [06a0599]
+  - @elbwalker/utils@3.1.0
+  - @elbwalker/client-node@1.0.1
+
 ## 1.0.1-next.0
 
 ### Patch Changes