+++ packages/types/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @elbwalker/types
 
+## 1.1.0
+
+### Minor Changes
+
+- d17e25a: walker on command added
+  [#314](https://github.com/elbwalker/walkerOS/issues/314)
+
+  Assign instance to window via config
+  [#316](https://github.com/elbwalker/walkerOS/issues/316)
+
+- 06a0599: on start command
+  [#325](https://github.com/elbwalker/walkerOS/issues/325)
+
 ## 1.1.0-next.0
 
 ### Minor Changes